requirements.txt
Django==1.8.5
djangorestframework==3.2.4
drf-nested-routers==0.10.0
httpie==0.9.2
psycopg2==2.6.1
Pygments==2.0.2
requests==2.8.0
